Fix React Routing on App.js and index.js, create header component and add it to all pages
