+++ src/pages/home/Home.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import Header from '../../components/header/Header'
+
+const Home = () => {
+    return (
+        <div>
+            
+        </div>
+    )
+}
+
+export default Home
