+++ src/pages/login/LoginPage.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react'
 import { toast, ToastContainer } from 'react-toastify';
-import { Link, useNavigate } from 'react-router-dom';
+import { Link } from 'react-router-dom';
